requirements_python3.9.txt
appscript==1.2.4
arabic-reshaper==2.1.3
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
Babel==2.9.1
backports.zoneinfo==0.2.1
build==0.9.0
bureau==0.0.1
certifi==2023.7.22
cffi==1.15.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect2==0.6.0
Cython==0.29.28
datetime-truncate==1.1.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-active-link==0.1.8
django-admin-charts==1.3.0
django-admin-interface==0.19.0
django-admin-relation-links==0.2.5
django-admin-tools==0.9.2
django-ajax-datatable==4.4.5
django-baton==2.3.0
django-bower==5.2.0
django-browser-reload==1.7.0
django-channels==0.7.0
django-ckeditor==6.5.1
django-clearcache==1.2.1
django-colorfield==0.7.0
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-crum==0.7.9
django-csp-plus==0.9.3
django-daterange-filterspec==2.0.4
django-dump-die==0.1.5
django-enumfield==3.1
django-environ==0.10.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==3.0.3
django-import-export==3.1.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-json-widget==1.1.1
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nested-inlines==0.1
django-nonrelated-inlines==0.1.1
django-nvd3==0.9.7
django-pandas==0.6.6
django-qr-code==3.1.1
django-qrcode==0.3
django-relativedelta==1.1.2
django-rest-swagger==2.2.0
django-session-timeout==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
et-xmlfile==1.1.0
fpdf==1.7.2
future==0.18.3
html5lib==1.1
idna==3.4
image==1.5.33
import-export==0.2.67.dev6
importlib-metadata==5.0.0
ipapi==1.0.4
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
Js2Py==0.74
lxml==4.8.0
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.3
marshmallow==3.15.0
msgpack==1.0.5
#mysqlclient==2.2.4
ndg-httpsclient==0.5.1
networkx==2.6.3
nltk==3.7
num2words==0.5.12
numpy==1.22.3
oauthlib==3.2.2
odfpy==1.4.1
onesignal-python-api==2.0.0
openapi-codec==1.3.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==21.3
panda==0.3.1
pandas==1.4.3
pep517==0.13.0
Pillow==9.5.0
pip-tools==6.10.0
polib==1.1.1
psutil==5.9.5
pusher==3.3.2
pyasn1==0.4.8
pycel==1.0b30
pycparser==2.21
pydantic==1.10.4
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyjsparser==2.7.1
PyJWT==2.6.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyotp==2.9.0
pyparsing==3.0.9
PyPDF2==3.0.1
PyPDF3==1.0.6
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-environ==0.4.54
python-gettext==5.0
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.5.4
regex==2022.6.2
register==0.1
reportlab==3.6.10
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
segno==1.5.2
self==2020.12.3
simplejson==3.19.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
svglib==1.3.0
tablib==3.3.0
tinycss2==1.1.1
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.4.0
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.6
uritemplate==4.1.1
uritools==4.0.0
urllib3==2.0.2
uWSGI==2.0.22
webencodings==0.5.1
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwings==0.28.5
xlwt==1.3.0
zipp==3.9.0
